+++ src/test/java/com/github/blutorange/maven/plugin/closurecompiler/test/MinifyMojoTest.java
@@ -13,6 +13,8 @@ import java.util.function.Function;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
+import com.github.blutorange.maven.plugin.closurecompiler.common.FileHelper;
+
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -22,8 +24,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-
-import com.github.blutorange.maven.plugin.closurecompiler.common.FileHelper;
 
 public class MinifyMojoTest {
 
@@ -66,6 +66,11 @@ public class MinifyMojoTest {
   @Test
   public void testSourceMap() throws Exception {
     runMinify("sourcemap");
+  }
+
+  @Test
+  public void testSubdirs() throws Exception {
+    runMinify("subdirs");
   }
 
   private void runMinify(String projectName) throws Exception {